Return payload in axios requests
