+++ src/actions/availabilities.js
@@ -1,19 +1,24 @@
 import axios from "axios";
 import urls from "../utils/urls";
 
-export const getAvailabilities = async () => axios.get(urls.apis.availability);
+export const getAvailabilities = async () =>
+  axios.get(urls.apis.availability).then((res) => res.data.payload);
 
 export const getAvailability = async (id) =>
-  axios.get(urls.apis.availability, {
-    params: {
-      id,
-    },
-  });
+  axios
+    .get(urls.apis.availability, {
+      params: {
+        id,
+      },
+    })
+    .then((res) => res.data.payload);
 
 export const addAvailability = async (availability) =>
-  axios.post(urls.apis.availability, {
-    availability,
-  });
+  axios
+    .post(urls.apis.availability, {
+      availability,
+    })
+    .then((res) => res.data.payload);
 
 export const deleteAvailability = async (id) =>
   axios.delete(urls.apis.availability, {
@@ -23,7 +28,9 @@ export const deleteAvailability = async (id) =>
   });
 
 export const updateAvailability = async (id, updatedFields) =>
-  axios.patch(urls.apis.availability, {
-    id,
-    updatedFields,
-  });
+  axios
+    .patch(urls.apis.availability, {
+      id,
+      updatedFields,
+    })
+    .then((res) => res.data.payload);
