+++ src/utils/axios.js
@@ -1,0 +1,20 @@
+import axios from "axios";
+
+axios.defaults.timeout = 15000;
+
+axios.interceptors.response.use(
+  function (response) {
+    if (response?.data && !response.data?.success) {
+      throw new Error(response.data.message);
+    }
+
+    return response;
+  },
+  function (error) {
+    if (error.response?.data && !error.response.data?.success) {
+      throw new Error(error.response.data.message);
+    }
+
+    throw error;
+  }
+);
